Fixed cycle bug and improved controls and minor improvements
Dumped PWM
Added functionaly to all buttons
Added invert key
SerialConn renamed and now detects based on os
Fixed cycle bug in usbConn

+++ arduino/usbConn.h
@@ -43,7 +43,7 @@ class usbConn{
         bool enableSendAck = true;
         char* buffer = 0;
         int bufferLoc = 0;
-        int timer = 0;
+        unsigned long timer = 0;
         int timeout = 1000;
         char checksum = 0;
         enum class readState{notReading, readCon, reading};
